requirements.txt
africastalking==1.2.9
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
ari==0.1.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
edge-tts==7.0.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.3.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-texttospeech==2.25.1
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Levenshtein==0.27.1
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.2.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
orjson==3.10.15
prometheus_client==0.21.1
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
python-dotenv==1.0.1
python-ESL==1.4.18
python-multipart==0.0.20
PyYAML==6.0.2
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
schema==0.7.7
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.39
srt==3.5.3
starlette==0.46.1
swaggerpy==0.2.1
sympy==1.13.1
tabulate==0.9.0
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
webrtcvad==2.0.10
websocket-client==1.8.0
websockets==15.0.1
yarl==1.18.3
